Rename various methods and clean up LSP

+++ lsp/src/main/java/vdmj/LSPDefinitionFinder.java
@@ -188,9 +188,9 @@ public class LSPDefinitionFinder
 		return null;
 	}
 	
-	public TCDefinition findDefinition(TCModuleList modules, File file, int line, int col)
+	public TCDefinition findDefinition(TCModuleList modules, File file, long line, long col)
 	{
-		LexLocation position = new LexLocation(file, "?", line, col, line, col);
+		LexLocation position = new LexLocation(file, "?", (int)line, (int)col, (int)line, (int)col);
 		Found found = findLocation(modules, position);
 		
 		if (found != null)
@@ -213,9 +213,9 @@ public class LSPDefinitionFinder
 		}
 	}
 	
-	public TCDefinition findDefinition(TCClassList classes, File file, int line, int col)
+	public TCDefinition findDefinition(TCClassList classes, File file, long line, long col)
 	{
-		LexLocation position = new LexLocation(file, "?", line, col, line, col);
+		LexLocation position = new LexLocation(file, "?", (int)line, (int)col, (int)line, (int)col);
 		Found found = findLocation(classes, position);
 		
 		if (found != null)
